work

+++ include/motor_action_data.h
@@ -2,6 +2,6 @@
 
 typedef struct {
     int motor_id;
-    int max_speed;
+    float max_speed;
 } motor_action_data_t;
